requirements.txt
structlog
pyyaml
